% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{analysis_plots}
\alias{analysis_plots}
\title{Create Analysis Plots}
\usage{
analysis_plots(rf.df.1, target, id_col, risk_col, time_col, vars_list)
}
\arguments{
\item{rf.df.1}{the dataframe}

\item{target}{the column name of the target variable}

\item{time_col}{the column name with the cpiu values}

\item{vars_list}{the variables used in training the random forest}

\item{id_Col}{the column name of the patient id variable}
}
\value{
shows the calibration and rpart plots for the model
}
\description{
\code{analysis_plots} creates plots of both the calibration of the random forest and the rpart summary tree
}
