% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{auc.smooth}
\alias{auc.smooth}
\title{Smooth AUC Values}
\usage{
auc.smooth(data, auc, target, patient_count_col)
}
\arguments{
\item{data}{The dataframe with the target values}

\item{auc}{The dataframe with the initial AUC values}

\item{target}{String variable containing the name of the target}

\item{patient_count_col}{String variable containing the name of the column couting the patient number}
}
\value{
a dataframe with a row for each CPIU and its smoothed AUC value
}
\description{
\code{auc.smooth} takes in a set of time varying AUC values and smooths them out, in addition to providing variance values
}
