% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{auc.smooth.return.single}
\alias{auc.smooth.return.single}
\title{Weighted Smooth AUC Value}
\usage{
auc.smooth.return.single(data, auc, target, patient_count_col)
}
\arguments{
\item{data}{The dataframe with the target values}

\item{auc}{The dataframe with the initial AUC values}

\item{target}{String variable containing the name of the target}

\item{patient_count_col}{String variable containing the name of the column couting the patient number}
}
\value{
a weighted average of the time varying CPIU values using weighting based on the number of patients available in each CPIU
}
\description{
\code{auc.smooth.return.single} both smooths the AUC values and then provides a single weighted average AUC based on the number
of observations used for the AUC value at each time
}
