% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p1.R
\name{boot.samp}
\alias{boot.samp}
\title{Bootstrap Sampling for RFSLAM}
\usage{
boot.samp(ntree = 100, id, boot.df.rep, id_col, time_unit_col)
}
\arguments{
\item{ntree}{the number of trees to sample for}

\item{id}{the list of patient ids}

\item{boot.df.rep}{the dataframe to bootstrap from}

\item{id_col}{the variable name holding the patient ids}

\item{time_unit_col}{the variable name holding the CPIU value}
}
\description{
\code{boot.samp} implements bootstrapping for RFSLAM
}
