% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{calc_risk_times}
\alias{calc_risk_times}
\title{Calculate Risk Times}
\usage{
calc_risk_times(
  df,
  end_time_col,
  curr_time_col,
  window_length,
  window_normalizer,
  col_name
)
}
\arguments{
\item{df}{the input dataframe}

\item{end_time_col}{column containing the final time value a patient had data for}

\item{curr_time_col}{column containing the time value that the row of data was collected}

\item{window_length}{lenghts of the CPIU intervals in the model}

\item{window_normalizer}{value to divide by}

\item{col_name}{name you want the new column created to have}
}
\description{
\code{calc_risk_times} calculates the risk times for each patient based on the given window lengths
}
