% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{create_model}
\alias{create_model}
\title{Train an RFSLAM model}
\usage{
create_model(
  modeling_df,
  target,
  id_col,
  risk_time_col,
  patient_count_col = "int.n",
  time_col,
  ntree = 100,
  nodedepth = NULL,
  nsplit = 10
)
}
\arguments{
\item{modeling_df}{dataframe containing data for modeling}

\item{target}{name of the target variable column}

\item{id_col}{name of the column with the patient id's}

\item{risk_time_col}{name of the column with the risk time values}

\item{patient_count_col}{name of the column with the patient counts}

\item{ntree}{number of trees for random forest}

\item{nodedepth}{the node depth for random forest}

\item{nsplit}{the nsplit parameter for random forest}
}
\value{
the RF-SLAM model, the predicted risk values (uncalibrated) together in a list that you can index into using the \code{model} and \code{preds} arguments
}
\description{
\code{create_model} trains a RFSLAM model using the given user parameter values
}
