% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{feature_importance_plot}
\alias{feature_importance_plot}
\title{Create Feature Importance Plot}
\usage{
feature_importance_plot(mymodel.1.full, var_key, importance_threshold = 10)
}
\arguments{
\item{mymodel.1.full}{the random forest model}

\item{var_key}{a variable key dataframe linking the variable names and their true meaning}

\item{importance_threshold}{minimum percent of trees variable should be in in order to be included in the plot}
}
\value{
shows the feature importance plot for the model
}
\description{
\code{feature_importance_plot} Plots the feature importance for the random forest
}
