% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{rf.auc}
\alias{rf.auc}
\title{Calculate time varying AUC dataframe}
\usage{
rf.auc(sca1.df = data, target, patient_count_col, time_col)
}
\arguments{
\item{sca1.df}{the data}

\item{target}{String variable containing the name of the target}

\item{patient_count_col}{String variable containing the name of the column couting the patient number}

\item{time_col}{String variable containing the name of the column holding the CPIU number}
}
\value{
time varying AUC dataframe with one column for the CPIU and another for the AUC
}
\description{
\code{rf.auc} calculates the time varying AUC values and returns them in a dataframe
}
