% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfsrc.news.R
\name{rfsrc.news}
\alias{rfsrc.news}
\title{Show the NEWS file}
\usage{
rfsrc.news(...)
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}
}
\value{
None.
}
\description{
Show the NEWS file of the \pkg{randomForestSRC} package.
}
\author{
Hemant Ishwaran and Udaya B. Kogalur
}
\keyword{documentation}
