% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{risk.adjust.new.be}
\alias{risk.adjust.new.be}
\title{Predicting Event Rates for New Data}
\usage{
risk.adjust.new.be(rf, status, rt, new_data, k = 2, alpha.tm = 0.05)
}
\arguments{
\item{rf}{the random forest object}

\item{status}{the target variable corresponding to "training" CPIUs used to build the rf}

\item{rt}{the risk time values corresponding to "training" CPIUs used to build the rf}

\item{new_data}{the new test dataset}
}
\value{
vector of risk values for the new data
}
\description{
\code{risk.adjust.new.be} calculates risk event rates on a new set of test data
}
