% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{save_risk_vals}
\alias{save_risk_vals}
\title{Save Risk Values}
\usage{
save_risk_vals(df, file_path)
}
\arguments{
\item{df}{the dataframe with the risk values}
}
\description{
\code{save_risk_vals} saves the risk values in \code{df} to the desired \code{file_path}
}
