% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{tune_rf_params}
\alias{tune_rf_params}
\title{Tune Parameters for RFSLAM}
\usage{
tune_rf_params(
  df,
  target,
  id_col,
  risk_time_col,
  patient_count_col = "int.n",
  time_col,
  drop,
  ntree_trys = c(50, 100, 200, 500),
  nodedepth_trys = c(NULL, 3, 5),
  nsplit_trys = c(5, 10, 15),
  n.folds = 5,
  folds_stratifier
)
}
\arguments{
\item{df}{the dataframe for modeling}

\item{target}{name of the target variable column}

\item{id_col}{name of the column with the patient id's}

\item{risk_time_col}{name of the column with the risk time values}

\item{patient_count_col}{name of the column with the patient counts}

\item{drop}{the variables to drop from the dataframe before training the model}

\item{ntree_trys}{the ntree options to try}

\item{nodedepth_trys}{the nodedepth options to try}

\item{nsplit_trys}{the nsplit options to try}

\item{n.folds}{the number of folds to use for cross validation}

\item{folds_stratifier}{the variable to use for creating the folds}
}
\value{
a table showing all combinations of the parameter values and their associated cross validated weighted auc value
}
\description{
\code{tune_rf_params} reports model auc values for parameter values searched over a grid as defined by the passed in parameters
}
